% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshist.R
\name{print.sshist}
\alias{print.sshist}
\title{Print method for sshist objects}
\usage{
\method{print}{sshist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sshist"}.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for sshist objects
}
