% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshist.R
\name{print.sshist_2d}
\alias{print.sshist_2d}
\title{Print method for sshist_2d objects}
\usage{
\method{print}{sshist_2d}(x, ...)
}
\arguments{
\item{x}{An object of class sshist.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print method for sshist_2d objects
}
