% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sskernel.R
\name{print.sskernel}
\alias{print.sskernel}
\title{Print method for sskernel objects}
\usage{
\method{print}{sskernel}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sskernel"}.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for sskernel objects
}
