% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sskernel.R
\name{print.sskernel2d}
\alias{print.sskernel2d}
\title{Print method for sskernel2d objects}
\usage{
\method{print}{sskernel2d}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sskernel2d"}.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for sskernel2d objects
}
