% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvkernel.R
\name{print.ssvkernel}
\alias{print.ssvkernel}
\title{Print method for ssvkernel objects}
\usage{
\method{print}{ssvkernel}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ssvkernel"}.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for ssvkernel objects
}
