% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvkernel.R
\name{print.ssvkernel2d}
\alias{print.ssvkernel2d}
\title{Print method for ssvkernel2d objects}
\usage{
\method{print}{ssvkernel2d}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ssvkernel2d"}.}

\item{...}{Additional arguments passed to \code{\link[base]{print}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for ssvkernel2d objects
}
