% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshist-package.R
\docType{package}
\name{sshist-package}
\alias{sshist-package}
\title{sshist: Optimal Density Estimation via Shimazaki-Shinomoto Method}
\description{
Implements the Shimazaki-Shinomoto method for optimizing the bin width of histograms and the bandwidth of kernel density estimators. The framework minimizes the expected Mean Integrated Squared Error (MISE) and supports both 1D and 2D distributions, fixed and locally adaptive estimators, bootstrap confidence intervals, and 'OpenMP'-accelerated 'C++' 'backends'. Ideally suited for time-dependent rate estimation and identifying intrinsic data structures. For more details see Shimazaki and Shinomoto (2007) \doi{10.1162/neco.2007.19.6.1503} and Shimazaki and Shinomoto (2010) \doi{10.1007/s10827-009-0180-4}.
}
\section{Main functions}{

\describe{
\item{\code{\link{sshist}}}{Optimal 1D histogram binning.}
\item{\code{\link{sshist_2d}}}{Optimal 2D histogram binning.}
\item{\code{\link{sskernel}}}{Optimal 1D fixed-bandwidth kernel density estimation.}
\item{\code{\link{sskernel2d}}}{Optimal 2D fixed-bandwidth kernel density estimation.}
\item{\code{\link{ssvkernel}}}{Locally adaptive 1D kernel density estimation.}
\item{\code{\link{ssvkernel2d}}}{Locally adaptive 2D kernel density estimation.}
}
}

\section{References}{

\itemize{
\item Shimazaki, H. and Shinomoto, S. (2007). A method for selecting the
bin size of a time histogram. \emph{Neural Computation}, \strong{19}(6),
1503–1527. \doi{10.1162/neco.2007.19.6.1503}
\item Shimazaki, H. and Shinomoto, S. (2010). Kernel bandwidth optimization
in spike rate estimation. \emph{Journal of Computational Neuroscience},
\strong{29}(1-2), 171–182. \doi{10.1007/s10827-009-0180-4}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/celebithil/sshist}
  \item \url{https://www.neuralengine.org/res/histogram.html}
  \item Report bugs at \url{https://github.com/celebithil/sshist/issues}
}

}
\author{
\strong{Maintainer}: Daniil Popov \email{popov.daniil@gmail.com}

Authors:
\itemize{
  \item Daniil Popov \email{popov.daniil@gmail.com}
}

}
\keyword{internal}
