% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshist.R
\name{plot.sshist}
\alias{plot.sshist}
\title{Plot method for sshist objects}
\usage{
\method{plot}{sshist}(x, ...)
}
\arguments{
\item{x}{An object of class sshist.}

\item{...}{Additional arguments passed to plot.}
}
\value{
No return value, called for side effects. Creates a two-panel plot showing:
(1) the cost function curve with the optimal number of bins highlighted, and
(2) the optimal histogram with the selected bin edges.
}
\description{
Plot method for sshist objects
}
