% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshist.R
\name{plot.sshist}
\alias{plot.sshist}
\title{Plot method for sshist objects}
\usage{
\method{plot}{sshist}(x, ...)
}
\arguments{
\item{x}{An object of class sshist.}

\item{...}{Additional arguments passed to plot.}
}
\description{
Plot method for sshist objects
}
