% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshist.R
\name{plot.sshist_2d}
\alias{plot.sshist_2d}
\title{Plot method for sshist_2d objects}
\usage{
\method{plot}{sshist_2d}(x, ...)
}
\arguments{
\item{x}{An object of class sshist.}

\item{...}{Additional arguments passed to plot.}
}
\description{
Plot method for sshist_2d objects
}
