% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshist.R
\name{plot.sshist_2d}
\alias{plot.sshist_2d}
\title{Plot method for sshist_2d objects}
\usage{
\method{plot}{sshist_2d}(x, ...)
}
\arguments{
\item{x}{An object of class sshist.}

\item{...}{Additional arguments passed to plot.}
}
\value{
No return value, called for side effects. Creates a two-panel plot showing:
(1) a heatmap of the cost function landscape across different bin combinations
with the optimal point highlighted in red, and (2) the optimal 2D histogram
visualization using the selected bin numbers for both dimensions.
}
\description{
Plot method for sshist_2d objects
}
